datasource-api-google-search-console/gsc-to-bq-cloud-function-trigger-one-url/requirements.txt

requests==2.25.1
pandas==1.5.2
pandas-gbq==0.19.0
python-dateutil==2.8.2
google-api-python-client==2.62.0